relocate swagger to new folder
